refactor(domain): centralize error types
