+++ backend/internal/domain/shared/error_types.go
@@ -1,0 +1,35 @@
+package shared
+
+import "fmt"
+
+// ValidationError represents a validation error with field information
+// Shared across domain packages.
+type ValidationError struct {
+	Field   string `json:"field"`
+	Message string `json:"message"`
+}
+
+func (e ValidationError) Error() string {
+	return fmt.Sprintf("validation error on field '%s': %s", e.Field, e.Message)
+}
+
+// NewValidationError creates a new validation error
+func NewValidationError(field, message string) ValidationError {
+	return ValidationError{Field: field, Message: message}
+}
+
+// BusinessRuleError represents a business rule violation
+// Shared across domain packages.
+type BusinessRuleError struct {
+	Rule    string `json:"rule"`
+	Message string `json:"message"`
+}
+
+func (e BusinessRuleError) Error() string {
+	return fmt.Sprintf("business rule violation '%s': %s", e.Rule, e.Message)
+}
+
+// NewBusinessRuleError creates a new business rule error
+func NewBusinessRuleError(rule, message string) BusinessRuleError {
+	return BusinessRuleError{Rule: rule, Message: message}
+}
